requirements.txt
streamlit
openai
beautifulsoup4
selenium
pdfplumber
pdfminer.six
python-pptx
openpyxl
pandas
lxml
